Base Commit 0.2
I fixed the entities and the DTO and added some lines of code to the repositories. The services are still empty.

TODO: figure out how to make the entities work accordingly to their foreign keys.  Having this many ID variables makes no sense, we could just use the entities as objects but I'm not sure how to do it best.